Updated readme
Reworked the text to use 'block' instead of 'message' for an encoded RS chunk of data; this is to prevent implying that there's a total limit on the size of the data you can store. The only limits are how you are forced to chunk apart your data into blocks, and then reading and writing any number of blocks to store your entire message.

Also fixed bad sentence explaining how messages are generated.